requirements.txt
appdirs==1.4.4
attrs==19.3.0
black==19.10b0
certifi==2020.6.20
click==7.1.2
cycler==0.10.0
importlib-metadata==1.7.0
iniconfig==1.0.1
kiwisolver==1.2.0
matplotlib==3.3.1
more-itertools==8.4.0
numpy==1.19.1
packaging==20.4
pathspec==0.8.0
Pillow==7.2.0
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pyrtlsdr==0.2.92
pytest==6.0.1
python-dateutil==2.8.1
PyYAML==5.3.1
regex==2020.7.14
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
zipp==3.1.0
